+++ tests/test_affine_transform.py
@@ -1,0 +1,161 @@
+import pyclesperanto_prototype as cle
+import numpy as np
+
+def test_affine_transform_translate():
+    source = cle.push_zyx(np.asarray([[
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+          [0, 0, 1, 1, 0],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+    ]]))
+
+    reference = cle.push_zyx(np.asarray([[
+        [0, 0, 0, 0, 0],
+        [0, 1, 1, 0, 0],
+        [0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0],
+    ]]))
+
+    transform = cle.AffineTransform3D()
+    transform.translate(-1, -1, 0)
+
+    result = cle.affine_transform(source, transform=transform)
+
+    a = cle.pull_zyx(result)
+    b = cle.pull_zyx(reference)
+
+    print(a)
+    print(b)
+
+    assert (np.array_equal(a, b))
+
+
+def test_affine_transform_scale():
+    source = cle.push_zyx(np.asarray([[
+          [0, 0, 0, 0, 0],
+          [0, 0, 1, 1, 0],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+    ]]))
+
+    reference = cle.push_zyx(np.asarray([[
+        [0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0],
+        [0, 0, 1, 1, 0],
+        [0, 0, 1, 1, 0],
+        [0, 0, 0, 0, 0],
+    ]]))
+
+    transform = cle.AffineTransform3D()
+    transform.scale(1, 2, 1)
+
+    result = cle.affine_transform(source, transform=transform)
+
+    a = cle.pull_zyx(result)
+    b = cle.pull_zyx(reference)
+
+    print(a)
+    print(b)
+
+    assert (np.array_equal(a, b))
+
+def test_affine_transform_scale_with_transform_matrix():
+    source = cle.push_zyx(np.asarray([[
+          [0, 0, 0, 0, 0],
+          [0, 0, 1, 1, 0],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+    ]]))
+
+    reference = cle.push_zyx(np.asarray([[
+        [0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0],
+        [0, 0, 1, 1, 0],
+        [0, 0, 1, 1, 0],
+        [0, 0, 0, 0, 0],
+    ]]))
+
+    transform = np.asarray([
+        [1, 0, 0, 0],
+        [0, 2, 0, 0],
+        [0, 0, 1, 0],
+        [0, 0, 0, 1],
+    ])
+
+    result = cle.affine_transform(source, transform=transform)
+
+    a = cle.pull_zyx(result)
+    b = cle.pull_zyx(reference)
+
+    print(a)
+    print(b)
+
+    assert (np.array_equal(a, b))
+
+
+def test_affine_transform_rotate():
+    source = cle.push_zyx(np.asarray([[
+          [0, 0, 0, 1, 1],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+    ]]))
+
+    reference = cle.push_zyx(np.asarray([[
+        [0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0],
+        [0, 0, 1, 0, 0],
+        [0, 0, 1, 1, 0],
+        [0, 0, 0, 0, 0],
+    ]]))
+
+    transform = cle.AffineTransform3D()
+    transform.rotate(2, 45.0)
+
+    result = cle.affine_transform(source, transform=transform)
+
+    a = cle.pull_zyx(result)
+    b = cle.pull_zyx(reference)
+
+    print(a)
+    print(b)
+
+    assert (np.array_equal(a, b))
+
+def test_affine_transform_rotate_around_center():
+    source = cle.push_zyx(np.asarray([[
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 1, 1],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+    ]]))
+
+    reference = cle.push_zyx(np.asarray([[
+        [0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0],
+        [0, 0, 1, 0, 0],
+        [0, 0, 1, 0, 0],
+    ]]))
+
+    transform = cle.AffineTransform3D()
+    transform.translate(-2.5, -2.5)
+    transform.rotate(2, 90.0)
+    transform.translate(2.5, 2.5)
+
+    result = cle.affine_transform(source, transform=transform)
+
+    a = cle.pull_zyx(result)
+    b = cle.pull_zyx(reference)
+
+    print(a)
+    print(b)
+
+    assert (np.array_equal(a, b))
+