+++ tests/test_rotate.py
@@ -1,0 +1,57 @@
+import pyclesperanto_prototype as cle
+import numpy as np
+
+def test_affine_transform_rotate():
+    source = cle.push_zyx(np.asarray([[
+          [0, 0, 0, 1, 1],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+    ]]))
+
+    reference = cle.push_zyx(np.asarray([[
+        [0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0],
+        [0, 0, 1, 0, 0],
+        [0, 0, 1, 1, 0],
+        [0, 0, 0, 0, 0],
+    ]]))
+
+    result = cle.rotate(source, angle_around_z_in_degrees=45.0, rotate_around_center=False)
+
+    a = cle.pull_zyx(result)
+    b = cle.pull_zyx(reference)
+
+    print(a)
+    print(b)
+
+    assert (np.array_equal(a, b))
+
+def test_affine_transform_rotate_around_center():
+    source = cle.push_zyx(np.asarray([[
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 1, 1],
+          [0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0],
+    ]]))
+
+    reference = cle.push_zyx(np.asarray([[
+        [0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0],
+        [0, 0, 1, 0, 0],
+        [0, 0, 1, 0, 0],
+    ]]))
+
+    result = cle.rotate(source, angle_around_z_in_degrees=90.0, rotate_around_center=True)
+
+    a = cle.pull_zyx(result)
+    b = cle.pull_zyx(reference)
+
+    print(a)
+    print(b)
+
+    assert (np.array_equal(a, b))
+